+++ lib/Register/authenticate.dart
