+++ lib/Screens/settings_menu.dart
@@ -35,7 +35,7 @@ class _SettingsMenuState extends State<Settings> {
             child: OutlineButton(
               onPressed: () {},
               child: Text(
-                'payment History',
+                'Payment History',
                 style: TextStyle(fontSize: 15, color: Colors.black),
               ),
             ),
